requirements.txt
tk
pyinstaller
pandas
requests
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
